refactor: Add CheckBoxFED widget for selecting fields
This commit adds a new file, checkBoxFED.dart, which contains the CheckBoxFED widget. This widget allows the user to select fields from a list, such as 'Frequência Cardíaca', 'SPO2', 'Pressão Arterial', 'Percepção Subjetiva de Esforço', and 'Dor'. The selected fields are then passed to the FieldsInicial page for further processing.

The CheckBoxFED widget is used in the paciente_page.dart file, where it is displayed when a patient's name is tapped. When the CheckBoxFED widget is tapped, it navigates to the FieldsInicial page with the selected fields and the patient data.

This commit also includes updates to the pubspec.lock file, updating the dependencies for bidi and ffi packages.